Add base adjoint ODE solver class
